Import using queues (#28)
* working queue worker

* various updates

corrected issues regarding imports.

* various changes

adjusted the grid layout in the data.blade.php to alow for 2 columns instead of only 3 or 1. Updated TableController  for importing. Moved some code into separate routines instead of having them in the process function.

* removed testing for php 5.6 in travis.

* fixed datacard issues

fixed overflow issues and added equal height js to keep the datacard at the same height. Also added a few items to views so the user can easily return to the record or table.

* updated controller test

since I adjusted the route the test was failing. modified the test so they pass.

* scss changes and import refactor

* comments added to tablecontroller.

* scrutinizer changes / removal of vagrant

* remove unneeded bootstrap.sh

* adjusted search.blade.php

adjusted blade to match data.blade.php since we changed scss and added javascript.

* using different method to determine delimiter

using different method to determine delimiter this only uses the first line to detect the delimiter.

* show changes

modified the show to detect some documents that were being showen in the in_text field. Wrote a routine to detect the filenames the filenames were .doc files but they have been converted to .txt do they have .doc.txt and are in the formletters folder.

* file detection

in the in_text field in the sample data we sometimes see files listed. This version is setup to look in formletters for any of these files since the path is incorrect also many files are already converted to .txt files.

* uncommented test